Fix/virtual queue with float value

+++ ros_queue/src/lib_queue/virtual_queue.cpp
@@ -3,7 +3,7 @@
 
 using std::invalid_argument;
 
-inline int VirtualQueue::size()
+inline float VirtualQueue::size()
 {
     return queue_size_;
 }
@@ -13,7 +13,7 @@ inline bool VirtualQueue::empty()
     return queue_size_ == 0;
 }
 
-inline void VirtualQueue::push(const int& nb_element)
+inline void VirtualQueue::push(const float& nb_element)
 {
     if(nb_element < 0)
     {
@@ -22,7 +22,7 @@ inline void VirtualQueue::push(const int& nb_element)
     queue_size_ += nb_element;
 }
 
-void VirtualQueue::pop(const int& nb_element)
+void VirtualQueue::pop(const float& nb_element)
 {
     if(nb_element < 0)
     {
@@ -39,7 +39,7 @@ void VirtualQueue::pop(const int& nb_element)
     }
 }
 
-inline void VirtualQueue::setSize(const int& new_size)
+inline void VirtualQueue::setSize(const float& new_size)
 {
     if(new_size < 0)
     {
@@ -49,17 +49,17 @@ inline void VirtualQueue::setSize(const int& new_size)
 }
 
 
-inline int NVirtualQueue::size()
+inline float NVirtualQueue::size()
 {
     return queue_size_;
 }
 
 inline bool NVirtualQueue::empty() 
 {
-    return queue_size_ == 0;
+    return queue_size_ == 0.0f;
 }
 
-inline void NVirtualQueue::push(const int& nb_element)
+inline void NVirtualQueue::push(const float& nb_element)
 {
     if(nb_element < 0)
     {
@@ -68,7 +68,7 @@ inline void NVirtualQueue::push(const int& nb_element)
     queue_size_ += nb_element;
 }
 
-inline void NVirtualQueue::pop(const int& nb_element)
+inline void NVirtualQueue::pop(const float& nb_element)
 {
     if(nb_element < 0)
     {
@@ -77,7 +77,7 @@ inline void NVirtualQueue::pop(const int& nb_element)
     queue_size_ -= nb_element;
 }
 
-inline void NVirtualQueue::setSize(const int& new_size)
+inline void NVirtualQueue::setSize(const float& new_size)
 {
     queue_size_ = new_size;
 }
